% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{Decode}
\alias{Decode}
\title{Decode}
\usage{
Decode(counts_file, map_file, params_file, alpha = 0.05,
  correction = c("Bonferroni"), quiet = FALSE, ...)
}
\arguments{
\item{counts_file}{Counts}

\item{map_file}{Map}

\item{params_file}{Params}

\item{alpha}{Alpha. Defaults to 0.05.}

\item{correction}{Correction. Defaults to Bonferroni.}

\item{quiet}{Quiet. Defaults to FALSE.}
}
\description{
This function decodes rappor reports.
}
\keyword{rappor}

